Fixes for signed / unsigned comparisons
I don't think these could manifest as bugs as the tree could never get
large enough in memory to overflow 63 bits. Found by compilation with
GCC 5.1 on F22.

+++ cpp/log/cluster_state_controller-inl.h
@@ -301,7 +301,7 @@ void ClusterStateController<Logged>::OnClusterStateUpdated(
       }
     } else {
       VLOG(1) << "Node left: " << node_id;
-      CHECK_EQ(1, all_peers_.erase(node_id));
+      CHECK_EQ(static_cast<size_t>(1), all_peers_.erase(node_id));
       fetcher_->RemovePeer(node_id);
     }
   }
